+++ src/main/java/u26c4/models/CalcResult.java
@@ -11,5 +11,5 @@ public class CalcResult {
     private String error;
     private String formattedExpression;
     private String rpn;
-    private Float calculation;
+    private Double calculation;
 }
